LIBTD-1270: based_near Location removed from the Item creation form; converted BatchUploadsController from class to override module and using prepend config/application.rb
